Bania_app/pyproject.toml
[project]
name = "bania"
version = "0.0.1"
description = "Microservicio Universal de IA para ecosistema clínico (Bania IA)"
authors = [
    { name = "BaniaIA", email = "sistemas@banpa.com" }
]
requires-python = ">=3.10"
dependencies = [
    "frappe",
    "google-generativeai"
]
